adding the docker file
