+++ pom.xml
@@ -25,6 +25,7 @@
 		<dependency>
 			<groupId>org.springframework.cloud</groupId>
 			<artifactId>spring-cloud-starter-netflix-eureka-client</artifactId>
+			<version>3.1.3</version>
 		</dependency>
 
 		<dependency>
