--- src/main/java/com/maveric/gatewayservice/configuration/GatewayConfiguration.java
@@ -1,61 +1,0 @@
-package com.maveric.gatewayservice.configuration;
-
-import com.maveric.gatewayservice.filter.AuthenticationPreFilter;
-//import com.maveric.gatewayservice.filter.JwtAuthFilter;
-import com.maveric.gatewayservice.properties.AccountProperties;
-import com.maveric.gatewayservice.properties.BalanceProperties;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.cloud.gateway.route.RouteLocator;
-import org.springframework.cloud.gateway.route.builder.RouteLocatorBuilder;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.cors.CorsConfiguration;
-import org.springframework.web.cors.reactive.CorsWebFilter;
-import org.springframework.web.cors.reactive.UrlBasedCorsConfigurationSource;
-
-import java.util.Arrays;
-import java.util.Collections;
-
-
-@Configuration
-public class GatewayConfiguration {
-
-    @Bean
-    public CorsWebFilter corsWebFilter() {
-
-        final CorsConfiguration corsConfig = new CorsConfiguration();
-        corsConfig.setAllowedOrigins(Collections.singletonList("*"));
-        corsConfig.setMaxAge(3600L);
-        corsConfig.setAllowedMethods(Arrays.asList("GET", "POST"));
-        corsConfig.addAllowedHeader("*");
-        corsConfig.addExposedHeader("Authorization");
-
-        final UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
-        source.registerCorsConfiguration("/api/v1/**", corsConfig);
-
-        return new CorsWebFilter(source);
-    }
-
-
-    @Bean
-    public RouteLocator gatewayRoutes(RouteLocatorBuilder builder, AuthenticationPreFilter authenticationPreFilter, BalanceProperties balanceProperties,
-                                      AccountProperties accountProperties) {
-        return builder.routes()
-
-                .route(routeToBalance -> routeToBalance.path(balanceProperties.getPath())
-                        .filters(f ->
-                                f.filter(authenticationPreFilter.apply(
-                                        new AuthenticationPreFilter.Config())))
-                        .uri(balanceProperties.getUri()))
-                .route(routeToAccount -> routeToAccount.path(accountProperties.getPath())
-                        .filters(f ->
-                                f.filter(authenticationPreFilter.apply(
-                                        new AuthenticationPreFilter.Config())))
-                        .uri(accountProperties.getUri()))
-
-
-                .build();
-    }
-    }
-
-
