+++ src/main/java/com/maveric/gatewayservice/properties/AuthProperties.java
@@ -1,0 +1,16 @@
+package com.maveric.gatewayservice.properties;
+
+import lombok.Getter;
+import lombok.Setter;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.stereotype.Component;
+
+@Getter
+@Setter
+@Component
+@ConfigurationProperties("auth-service")
+public class AuthProperties {
+    private String path;
+
+    private String uri;
+}
