+++ src/main/java/com/maveric/gatewayservice/properties/TransactionProperties.java
@@ -1,0 +1,15 @@
+package com.maveric.gatewayservice.properties;
+
+import lombok.Getter;
+import lombok.Setter;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.stereotype.Component;
+@Getter
+@Setter
+@Component
+@ConfigurationProperties("transaction-service")
+public class TransactionProperties {
+    private String path;
+
+    private String uri;
+}
