+++ src/main/resources/application.properties
@@ -9,10 +9,17 @@ spring.main.allow-bean-definition-overriding=true
 balance-service.path=/api/v1/accounts/{balanceId}/balances/**
 balance-service.uri=http://localhost:3015
 
-
+transaction-service.path=/api/v1/accounts/{accountId}/**
+transaction-service.uri=http://localhost:3020
 
 account-service.path=/api/v1/customers/{customerId}/**
 account-service.uri=http://localhost:3010
+
+auth-service.path=/api/v1/auth/**
+auth-service.uri=http://localhost:3000
+
+user-service.path=/api/v1/users/**
+user-service.uri=http://localhost:3005
 
 eureka.client.register-with-eureka=false
